rename: quote
