+++ controllers/quote.controller.js
@@ -1,0 +1,11 @@
+const Quote = require("../models/quote.model");
+const { StatusCodes } = require("http-status-codes");
+exports.getAllQuotes = async (req, res) => {
+  const getAllQuotes = await Quote.getAllQuotes();
+
+  const randomNumber = Math.floor(Math.random() * getAllQuotes.length);
+
+  const getOneQuote = await Quote.getOneQuote(randomNumber);
+
+  return res.status(StatusCodes.OK).json({ Quotes: getOneQuote });
+};
