+++ routes/quote.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const Quote = require("../controllers/quote");
+const Quote = require("../controllers/quote.controller");
 
 const asyncWrapper = require("../middleware/async");
